refactor: add BaseRuntime and BasePea
