--- jina/peapods/peas/gateway/__init__.py
@@ -1,5 +1,0 @@
-__copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
-__license__ = "Apache-2.0"
-
-from .grpc import GatewayPea
-from .rest import RESTGatewayPea
